Fixed type problem in TranspositionEntry

+++ src/main/java/za/ac/sun/cs/adversarial/transposition/TranspositionEntry.java
@@ -43,7 +43,7 @@ public class TranspositionEntry {
         return this.depth;
     }
 
-    public void setKey(int key) {
+    public void setKey(long key) {
         this.key = key;
     }
